autorandr, the eternal struggle
This commit may finally settle it. We now use redshift's one-shot manual
flag to fix the screen temperature no matter what time of the day it is.
And we save autorandr with the gamma options as it is supposed to be
consistent now.

This commit also restarts picom when switching autorandr as the switch
seems to kill it.