change default browser from zen to chrome
yeah..zen was consuming more RAM. and chrome
is faster. and i need to develop in chrome since
it is the most used browser by users and some
issues only happen in chrome. so, for simplicity
(stickying to 1 browser), i'm surrendering to chrome
(at least for now)

also, change the function webapp to a script so it can
be called by other scripts (don't know what was I thinking
when I decided to create a bash function lol)